Revert "Coding/trúc"
