+++ src/Admin/ManageSchedule/AddSchedule.js
@@ -104,7 +104,6 @@ export default function AddSchedule() {
         startTime,
         endTime,
         slotBooking: 1,
-        status: true,
       };
 
       const response = await fetch(SCHEDULE_API.MASTER, {
