+++ src/Components/Login/Login.js
@@ -84,8 +84,6 @@ export default function Login() {
         console.log("Role", decodedToken.Role === "Staff");
         if (decodedToken.Role === "Staff") {
           navigate("/manage-appointment");
-        } else if (decodedToken.Role === "Doctor") {
-          navigate("/doctor-schedule");
         } else {
           navigate("/");
         }
