+++ src/HomePage/Booking/Booking.js
@@ -37,35 +37,55 @@ export default function Booking() {
     if (user) {
       fetchUserPets();
       fetchDoctors();
+      fetchServices();
     } else {
       navigate("/signin"); // Redirect to login if user is not authenticated
     }
   }, [user, navigate]);
 
-  useEffect(() => {
-    if (selectedDoctorId) {
-      fetchServices(selectedDoctorId);
-    }
-  }, [selectedDoctorId]);
-
-  useEffect(() => {
-    if (selectedDoctorId || date) {
-      fetchAvailableSchedules(selectedDoctorId, date);
-    }
-  }, [selectedDoctorId, date]);
-
-  const fetchServices = async (doctorId) => {
+  const fetchAvailableSchedules = async (doctorId, selectedDate) => {
     try {
-      const response = await fetch(`${DOCTOR_API.MASTER}/${doctorId}`);
+      const response = await fetch(
+        `${SCHEDULE_API.MASTER}?DoctorId=${doctorId}`
+      );
+      if (!response.ok) {
+        throw new Error(`HTTP error! Status: ${response.status}`);
+      }
       const data = await response.json();
 
-      if (data.data.serviceList) {
-        setServices(data.data.serviceList);
+      // Filter schedules by selected date and doctorId
+      const filteredSchedules = data.filter((schedule) => {
+        // Assuming schedule.startTime is in ISO string format
+        const scheduleDate = new Date(schedule.startTime)
+          .toISOString()
+          .split("T")[0]; // Extract date part in ISO format
+        const selectedDateString = new Date(selectedDate)
+          .toISOString()
+          .split("T")[0]; // Convert selectedDate to ISO format
+
+        return scheduleDate === selectedDateString && schedule.status === true;
+      });
+
+      setAvailableSchedules(filteredSchedules);
+
+      if (filteredSchedules.length === 0) {
+        setSelectedSchedule(""); // Clear selected schedule if no schedules available
       } else {
-        setServices([]);
+        setSelectedSchedule(filteredSchedules[0].scheduleId); // Set default to first available schedule
       }
     } catch (error) {
+      console.error("Error fetching schedules:", error);
+      toast.error("Error fetching schedules");
+    }
+  };
+  const fetchServices = async () => {
+    try {
+      const response = await fetch(SERVICE_API.MASTER);
+      const data = await response.json();
+      setServices(data || []);
+    } catch (error) {
       console.error("Error fetching services:", error);
+      toast.error("Error fetching services");
     }
   };
 
@@ -76,12 +96,13 @@ export default function Booking() {
       setDoctors(data.data.items || []);
     } catch (error) {
       console.error("Error fetching doctors:", error);
-      // toast.error("Error fetching doctors");
+      toast.error("Error fetching doctors");
     }
   };
 
   const fetchUserPets = async () => {
     try {
+      console.log("userId", user.customerId);
       const response = await fetch(
         `${PET_API.MASTER}?CustomerId=${user.customerId}`
       );
@@ -90,57 +111,7 @@ export default function Booking() {
       setUserPets(data.data.items || []);
     } catch (error) {
       console.error("Error fetching user pets:", error);
-    }
-  };
-
-  const fetchAvailableSchedules = async (doctorId, selectedDate) => {
-    console.log("doctorId", doctorId);
-    try {
-      const response = await fetch(
-        `${SCHEDULE_API.MASTER}?DoctorId=${doctorId}`
-      );
-      if (!response.ok) {
-        throw new Error(`HTTP error! Status: ${response.status}`);
-      }
-      const data = await response.json();
-
-      console.log("data", data);
-
-      const filteredSchedules = data.filter((schedule) => {
-        const scheduleDate = new Date(schedule.startTime)
-          .toISOString()
-          .split("T")[0];
-        const selectedDateString = new Date(selectedDate)
-          .toISOString()
-          .split("T")[0];
-
-        return scheduleDate === selectedDateString && schedule.status === true;
-      });
-      console.log("filter schedule", filteredSchedules);
-
-      setAvailableSchedules(filteredSchedules);
-
-      if (filteredSchedules.length === 0) {
-        setSelectedSchedule("");
-      } else {
-        setSelectedSchedule(filteredSchedules[0].scheduleId);
-      }
-
-      const uniqueDoctorIds = Array.from(
-        new Set(filteredSchedules.map((schedule) => schedule.doctorId))
-      );
-
-      const doctorsData = await Promise.all(
-        uniqueDoctorIds.map(async (doctorId) => {
-          const response = await fetch(`${DOCTOR_API.MASTER}/${doctorId}`);
-          const data = await response.json();
-          return data.data;
-        })
-      );
-
-      setDoctors(doctorsData);
-    } catch (error) {
-      console.error("Error fetching schedules:", error);
+      // toast.error("Error fetching user pets");
     }
   };
 
@@ -161,7 +132,9 @@ export default function Booking() {
         petId: parseInt(selectedPetId, 10),
         customerId: user.customerId,
         doctorId: parseInt(selectedDoctorId, 10),
-        serviceIds: selectedServices,
+        serviceIds: selectedServices.map((serviceId) =>
+          parseInt(serviceId, 10)
+        ),
         note: note,
         scheduleId: parseInt(selectedSchedule, 10),
       };
@@ -213,13 +186,6 @@ export default function Booking() {
     }
   };
 
-  const getTodayDate = () => {
-    const today = new Date();
-    const year = today.getFullYear();
-    const month = String(today.getMonth() + 1).padStart(2, "0");
-    const day = String(today.getDate()).padStart(2, "0");
-    return `${year}-${month}-${day}`;
-  };
   return (
     <div>
       <ToastContainer />
@@ -384,22 +350,6 @@ export default function Booking() {
                         </div>
                         <div className="col-md-6">
                           <div className="mb-3">
-                            <label className="form-label" htmlFor="input-date">
-                              Date
-                            </label>
-                            <input
-                              name="date"
-                              type="date"
-                              className="form-control"
-                              id="input-date"
-                              value={date}
-                              min={getTodayDate()}
-                              onChange={handleDateChange}
-                            />
-                          </div>
-                        </div>
-                        <div className="col-md-6">
-                          <div className="mb-3">
                             <label
                               className="form-label"
                               htmlFor="select-doctor"
@@ -428,7 +378,21 @@ export default function Booking() {
                             </select>
                           </div>
                         </div>
-
+                        <div className="col-md-6">
+                          <div className="mb-3">
+                            <label className="form-label" htmlFor="input-date">
+                              Date
+                            </label>
+                            <input
+                              name="date"
+                              type="date"
+                              className="form-control"
+                              id="input-date"
+                              value={date}
+                              onChange={handleDateChange}
+                            />
+                          </div>
+                        </div>
                         <div className="col-md-6">
                           <div className="mb-3">
                             <label className="form-label">
